Move progress bar into a module
